Use MaxArea limitation.

+++ Rfunction/FLDAS_nc2RDS.R
@@ -5,14 +5,23 @@ require(ncdf4)
 source('GetReady.R')
 source('Rfunction/Fun.readnc.R')
 
+fixNA <- function(x){
+  for(i in 1:ncol(x)){
+    cv=x[,i]
+    cv[is.na(cv)] = mean(cv, na.rm=TRUE)
+    x[,i] = cv
+  }
+  x
+}
+
 fl=readOGR(file.path(dir.pihmgis, 'FLDAS_GCS.shp'))@data
 head(fl)
 # years=2017:2018
-dirs = file.path(dir.fldas, years)
+dirs = file.path(dir.ldas, years)
 
 ndir = length(dirs)
-fn=list.files(dir.fldas, pattern=glob2rx('*.nc'), recursive = T, full.names = T)[1]
-
+fn=list.files(dir.ldas, pattern=glob2rx('*.nc'), recursive = TRUE, full.names = T)[1]
+fn
 fid=nc_open(fn)
 xloc = round(fid$dim$X$vals,2)
 yloc = round(fid$dim$Y$vals, 2)
@@ -43,25 +52,32 @@ dev.off()
 
 nv=length(vns)
 for(idd  in 1:ndir){ # for each year dir
-# library(foreach)
-# library(doMC)
-# library(doParallel)
-# registerDoMC(4)
-# foreach (idd = 1:ndir) %dopar%{
+  # library(foreach)
+  # library(doMC)
+  # library(doParallel)
+  # registerDoMC(4)
+  # foreach (idd = 1:ndir) %dopar%{
   cdir <- dirs[idd]
-  fns = list.files(cdir, pattern=glob2rx('*.nc'), recursive = T, full.names = T)
-  nf = length(fns)
-  x.arr = array(0, dim=c(ns, nv, nf) )
-  x.t= character(nf)
-  for(j  in 1:nf){  # files in each year
-    fn=fns[j]
-    t=substr(basename(fn), 22, 29)
-    message(j, '/', nf, '\t', t)
-    x.mat = readnc(fn, xyid=xyid, vns=vns)
-    x.t[j] = t
-    x.arr[,,j ] = x.mat 
-  }
-  dimnames(x.arr) = list(sn, vns,  x.t)
   fn.rds = file.path(dir.pihmgis, paste0(prjname,'-', basename(cdir), '.RDS'))
-  saveRDS(x.arr, file=fn.rds)
+  
+  if(!file.exists(fn.rds)){
+    fns = list.files(cdir, pattern=glob2rx('*.nc'), recursive = T, full.names = T)
+    nf = length(fns)
+    x.arr = array(0, dim=c(ns, nv, nf) )
+    x.t= character(nf)
+    for(j  in 1:nf){  # files in each year
+      fn=fns[j]
+      t=substr(basename(fn), 22, 29)
+      message(j, '/', nf, '\t', t)
+      x.mat = readnc(fn, xyid=xyid, vns=vns)
+      x.mat = fixNA(x.mat)
+      x.t[j] = t
+      x.arr[,,j ] = x.mat 
+      if(sum(is.na(x.mat))>1){
+        message('NA value exist in data, at ', j)
+      }
+    }
+    dimnames(x.arr) = list(sn, vns,  x.t)
+    saveRDS(x.arr, file=fn.rds)
+  }
 }
